PLANNER-1652 Move TestdataSolutionStateRecorder to test class
